Replaced the thread pool refresh implementation with one using Actors
Details
-------

SwingMain.scala:

Changed the thread pool over to use two actors.  The first just sits there firing refresh messages off every 250ms, the other loops waiting for messages.  If all I needed was the refresh I could have managed with one actor, but since I watned to extend things to allow for some UI commands as well, I needed to avoid having the main actor block so the refresh part went off into a separate actor.

Added a couple of admitedly arbirary buttons to the UI, a Stop/Start one, that pauses the refresh (or disables/enables listening to the refresh message to be precise), and a Random one that reloads the board with a new random map.  The functionality is basic but it let me exersize passing messages to actors.

Replaces the swing references to size with preferredSize since that's actually what works, now the layout works as expected.  Scala Swing docs really suck, I guess you really need to know Swing for this to be clear.

Fixed some off by one issues in the board size in the UI.

Board.scala:

Added a Board.clear() function.  Changed the val grid to be a var.  Both of these are so that the grid can be set to null, I'm pretty sure that this is a scala bug, but without this, the reference to the grid appears to never get garbage collected.

Added preconditions to getLiveNeighborsCount().

switched greateGrid() to add to a mutable map as it goes along and to create a immutable one as it's return.  I did this while I was trying to track down the memory leak and after the fact decided that it seemed a better implemetation.  Probably just a matter of taste.

BoardFactory.scala:

Nothing to speak of, though I agree with the idea of switching it to be a Board object, I think that that is a tidier "scala" way of doing it.

BoardFactorySpec.scala:

Changed the board string to use the scala margin pattern.  Seems like a very neat way to clean up multi-line string constants.

BoardSpec.scala:

String constant change here too.

Added a few more tests, nothing major really though.

--------------------------------------------------------------------------------